# archive/root_legacy/artifacts/openapi.json
{"openapi":"3.1.0","info":{"title":"Ancient Text Analyzer","description":"Advanced OCR and cryptographic analysis system for historical texts","version":"1.0.0"},"paths":{"/api/v1/auth/login":{"post":{"tags":["auth"],"summary":"Login","description":"User login endpoint\nReturns JWT access token for authenticated requests","operationId":"login_api_v1_auth_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/register":{"post":{"tags":["auth"],"summary":"Register","description":"User registration endpoint\nCreates new user account (demo implementation)","operationId":"register_api_v1_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/me":{"get":{"tags":["auth"],"summary":"Get Current User Info","description":"Get current user information\nRequires valid authentication token","operationId":"get_current_user_info_api_v1_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/auth/logout":{"post":{"tags":["auth"],"summary":"Logout","description":"User logout endpoint\nIn a stateless JWT system, logout is handled client-side by discarding the token","operationId":"logout_api_v1_auth_logout_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/auth/refresh":{"post":{"tags":["auth"],"summary":"Refresh Token","description":"Refresh access token\nIssues a new token with extended expiration","operationId":"refresh_token_api_v1_auth_refresh_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/auth/demo-users":{"get":{"tags":["auth"],"summary":"Get Demo Users","description":"Get list of demo users for development/testing\nRemove this endpoint in production!","operationId":"get_demo_users_api_v1_auth_demo_users_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/patterns/{document_id}":{"get":{"tags":["patterns"],"summary":"Get Patterns","description":"Get patterns for a document with filtering","operationId":"get_patterns_api_v1_patterns__document_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"document_id","in":"path","required":true,"schema":{"type":"integer","title":"Document Id"}},{"name":"pattern_type","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Pattern Type"}},{"name":"min_confidence","in":"query","required":false,"schema":{"type":"number","default":0.0,"title":"Min Confidence"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PatternResponse"},"title":"Response Get Patterns Api V1 Patterns  Document Id  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/patterns/detail/{pattern_id}":{"get":{"tags":["patterns"],"summary":"Get Pattern Detail","description":"Get detailed information about a specific pattern","operationId":"get_pattern_detail_api_v1_patterns_detail__pattern_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"pattern_id","in":"path","required":true,"schema":{"type":"integer","title":"Pattern Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PatternResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/search/":{"post":{"tags":["search"],"summary":"Search","description":"Perform advanced search across documents and patterns","operationId":"search_api_v1_search__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SearchRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SearchResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"HTTPBearer":[]}]}},"/api/v1/search/suggestions":{"get":{"tags":["search"],"summary":"Get Search Suggestions","description":"Get search suggestions and autocomplete","operationId":"get_search_suggestions_api_v1_search_suggestions_get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"query","in":"query","required":true,"schema":{"type":"string","title":"Query"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/relationships/network":{"post":{"tags":["relationships"],"summary":"Get Relationship Network","description":"Build a network graph showing relationships between selected documents.\nUsed for visualization (node-link diagrams).","operationId":"get_relationship_network_api_v1_relationships_network_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"request_id","in":"query","required":false,"schema":{"type":"string","default":"default","title":"Request Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AnalysisRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RelationshipNetworkResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/relationships/matrix":{"post":{"tags":["relationships"],"summary":"Get Correlation Matrix","description":"Generate a correlation matrix (heatmap) for a set of documents.\nCompares signatures across pattern, stylistic, temporal, and geometric dimensions.","operationId":"get_correlation_matrix_api_v1_relationships_matrix_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"request_id","in":"query","required":false,"schema":{"type":"string","default":"default","title":"Request Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AnalysisRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CorrelationMatrixResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/relationships/authorship":{"post":{"tags":["relationships"],"summary":"Analyze Authorship","description":"Analyze authorship patterns to detect if documents were likely created by the same entity.","operationId":"analyze_authorship_api_v1_relationships_authorship_post","security":[{"HTTPBearer":[]}],"parameters":[{"name":"request_id","in":"query","required":false,"schema":{"type":"string","default":"default","title":"Request Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AnalysisRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthorshipProfileResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/relationships/trails/{pattern_id}":{"get":{"tags":["relationships"],"summary":"Track Evidence Trails","description":"Trace the 'Evidence Trail' of a specific pattern across multiple documents.\nShows how a specific symbol or cipher traverses the corpus.","operationId":"track_evidence_trails_api_v1_relationships_trails__pattern_id__get","security":[{"HTTPBearer":[]}],"parameters":[{"name":"pattern_id","in":"path","required":true,"schema":{"type":"integer","title":"Pattern Id"}},{"name":"request_id","in":"query","required":false,"schema":{"type":"string","default":"default","title":"Request Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/EvidenceTrailResponse"},"title":"Response Track Evidence Trails Api V1 Relationships Trails  Pattern Id  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","description":"Health check endpoint","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health Check","description":"Detailed health check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/upload":{"post":{"summary":"Upload File","description":"Upload a document for analysis","operationId":"upload_file_api_v1_upload_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_file_api_v1_upload_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AnalysisRequest":{"properties":{"document_ids":{"items":{"type":"integer"},"type":"array","title":"Document Ids"}},"type":"object","required":["document_ids"],"title":"AnalysisRequest"},"AuthorshipProfileResponse":{"properties":{"document_count":{"type":"integer","title":"Document Count"},"similarity_matrix":{"items":{"items":{"type":"number"},"type":"array"},"type":"array","title":"Similarity Matrix"},"author_groups":{"additionalProperties":{"items":{"type":"integer"},"type":"array"},"type":"object","title":"Author Groups"},"confidence_scores":{"additionalProperties":{"type":"number"},"type":"object","title":"Confidence Scores"},"analysis_summary":{"type":"string","title":"Analysis Summary"}},"type":"object","required":["document_count","similarity_matrix","author_groups","confidence_scores","analysis_summary"],"title":"AuthorshipProfileResponse"},"Body_upload_file_api_v1_upload_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_file_api_v1_upload_post"},"CorrelationMatrixResponse":{"properties":{"document_ids":{"items":{"type":"integer"},"type":"array","title":"Document Ids"},"matrix":{"items":{"items":{"type":"number"},"type":"array"},"type":"array","title":"Matrix"},"significance_matrix":{"items":{"items":{"type":"number"},"type":"array"},"type":"array","title":"Significance Matrix"},"cluster_assignments":{"additionalProperties":{"type":"integer"},"type":"object","title":"Cluster Assignments"}},"type":"object","required":["document_ids","matrix","significance_matrix","cluster_assignments"],"title":"CorrelationMatrixResponse"},"EvidencePoint":{"properties":{"type":{"type":"string","title":"Type"},"strength":{"type":"number","title":"Strength"},"details":{"type":"object","title":"Details"}},"type":"object","required":["type","strength","details"],"title":"EvidencePoint"},"EvidenceTrailResponse":{"properties":{"pattern_id":{"type":"integer","title":"Pattern Id"},"document_chain":{"items":{"type":"integer"},"type":"array","title":"Document Chain"},"connection_strength":{"type":"number","title":"Connection Strength"},"confidence_score":{"type":"number","title":"Confidence Score"},"evidence_points":{"items":{"$ref":"#/components/schemas/EvidencePoint"},"type":"array","title":"Evidence Points"}},"type":"object","required":["pattern_id","document_chain","connection_strength","confidence_score","evidence_points"],"title":"EvidenceTrailResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"LoginRequest":{"properties":{"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"}},"type":"object","required":["username","password"],"title":"LoginRequest"},"LoginResponse":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type"},"expires_in":{"type":"integer","title":"Expires In"},"user":{"type":"object","title":"User"}},"type":"object","required":["access_token","token_type","expires_in","user"],"title":"LoginResponse"},"NetworkEdge":{"properties":{"source":{"type":"integer","title":"Source"},"target":{"type":"integer","title":"Target"},"weight":{"type":"number","title":"Weight"},"relationship_type":{"type":"string","title":"Relationship Type"},"evidence_count":{"type":"integer","title":"Evidence Count"},"confidence":{"type":"number","title":"Confidence"}},"type":"object","required":["source","target","weight","relationship_type","evidence_count","confidence"],"title":"NetworkEdge"},"NetworkGraph":{"properties":{"nodes":{"items":{"$ref":"#/components/schemas/NetworkNode"},"type":"array","title":"Nodes"},"edges":{"items":{"$ref":"#/components/schemas/NetworkEdge"},"type":"array","title":"Edges"}},"type":"object","required":["nodes","edges"],"title":"NetworkGraph"},"NetworkNode":{"properties":{"id":{"type":"integer","title":"Id"},"title":{"type":"string","title":"Title"},"centrality":{"type":"number","title":"Centrality"},"community":{"type":"integer","title":"Community"}},"type":"object","required":["id","title","centrality","community"],"title":"NetworkNode"},"PatternResponse":{"properties":{"id":{"type":"integer","title":"Id"},"pattern_type":{"type":"string","title":"Pattern Type"},"description":{"type":"string","title":"Description"},"confidence":{"type":"number","title":"Confidence"},"page_number":{"type":"integer","title":"Page Number"},"coordinates":{"anyOf":[{"items":{"type":"object"},"type":"array"},{"type":"null"}],"title":"Coordinates"}},"type":"object","required":["id","pattern_type","description","confidence","page_number"],"title":"PatternResponse"},"RegisterRequest":{"properties":{"username":{"type":"string","title":"Username"},"email":{"type":"string","title":"Email"},"password":{"type":"string","title":"Password"}},"type":"object","required":["username","email","password"],"title":"RegisterRequest"},"RelationshipNetworkResponse":{"properties":{"network":{"$ref":"#/components/schemas/NetworkGraph"},"metrics":{"type":"object","title":"Metrics"},"communities":{"additionalProperties":{"type":"integer"},"type":"object","title":"Communities"},"central_documents":{"items":{"type":"integer"},"type":"array","title":"Central Documents"},"summary":{"type":"object","title":"Summary"}},"type":"object","required":["network","metrics","communities","central_documents","summary"],"title":"RelationshipNetworkResponse"},"SearchRequest":{"properties":{"query":{"type":"string","title":"Query"},"document_ids":{"anyOf":[{"items":{"type":"integer"},"type":"array"},{"type":"null"}],"title":"Document Ids"},"pattern_types":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Pattern Types"},"min_confidence":{"type":"number","title":"Min Confidence","default":0.0},"max_results":{"type":"integer","title":"Max Results","default":100}},"type":"object","required":["query"],"title":"SearchRequest"},"SearchResponse":{"properties":{"results":{"items":{"type":"object"},"type":"array","title":"Results"},"total_found":{"type":"integer","title":"Total Found"},"query_time":{"type":"number","title":"Query Time"}},"type":"object","required":["results","total_found","query_time"],"title":"SearchResponse"},"UserResponse":{"properties":{"username":{"type":"string","title":"Username"},"email":{"type":"string","title":"Email"},"is_active":{"type":"boolean","title":"Is Active"}},"type":"object","required":["username","email","is_active"],"title":"UserResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"HTTPBearer":{"type":"http","scheme":"bearer"}}}}